Working

+++ mc2/tools/minimize.py
@@ -1,23 +1,29 @@
-import sys, os, time
+import os
 from ..runner import *
 from ..maze_gen import smt2_parser as sp
 
-def main(maze_path,tool,variant, memory,mutant,outdir,timeout):
-    params = get_params(maze_path)
+def main(maze_path,tool,variant,seeddir, outdir,timeout):
+    params = get_params(maze_path,seeddir)
     commands.run_cmd('mkdir -p %s' % outdir)
     commands.run_cmd("mkdir -p %s" % os.path.join(outdir,'seeds'))
     commands.run_cmd("mkdir -p %s" % os.path.join(outdir,'runs'))
- 
-    clauses = get_clauses(mutant)
-    misses_bug = True
-    keep_first_half = True
 
+    if not check_if_tp(tool, variant, params, outdir, timeout):
+        print('ERROR: Original not maze not a fn')
+        return
+    mutant = os.path.join(outdir,'smt', 'mutant_%d.smt2' % (params['m'] - 1))
+    clauses = get_clauses(mutant)
+
+    keep_first_half = True
+    misses_bug = True 
+    params = set_fake_params(params)
     while misses_bug or not keep_first_half:
         half = len(clauses) // 2
         new_clauses = clauses[:half] if keep_first_half else clauses[half+1:]
 
         seed = os.path.join(outdir, 'seeds', str(half) + ('-first' if keep_first_half else '-second'))
-        misses_bug = check_result(tool, variant, memory, params, outdir, seed,new_clauses,timeout)
+        set_seed(params,seed,new_clauses)
+        misses_bug = check_if_tp(tool, variant, params, outdir,timeout)
 
         commands.run_cmd('rm -r %s' % os.path.join(outdir, 'src'))
         if misses_bug:
@@ -34,8 +40,9 @@ def main(maze_path,tool,variant, memory,mutant,outdir,timeout):
         pos = i - empty_clauses
         seed = os.path.join(outdir, 'seeds', str(len(clauses)) + '-' + str(pos+1))
         clause = clauses.pop(pos)
-        
-        misses_bug = check_result(tool,variant,memory, params, outdir, seed,clauses,timeout)
+    
+        set_seed(params,seed,clauses)
+        misses_bug = check_if_tp(tool,variant,params, outdir,timeout)
         commands.run_cmd('rm -r %s' % os.path.join(outdir, 'src'))
         
         if misses_bug:
@@ -45,12 +52,11 @@ def main(maze_path,tool,variant, memory,mutant,outdir,timeout):
         
         
     set_seed(params,seed,clauses)
-    maze_gen.generate_maze(params,out_dir = outdir)
+    maze_gen.generate_mazes([params],outdir)
         
 
-def check_result(tool,variant,memory, params, outdir, seed, clauses,timeout):
-    set_seed(params, seed, clauses)
-    docker.run_mc(tool,variant, 'min', memory, params, outdir,timeout)
+def check_if_tp(tool,variant,params, outdir, timeout):
+    docker.run_mc(tool,variant, 'min', params, outdir,timeout=timeout)
     sat = is_fn(outdir)
     return sat
 
@@ -75,9 +81,12 @@ def is_fn(outdir):
                     return True
     return False
 
-def get_params(maze_path):
+def get_params(maze_path, smt_path = ''):
     maze = maze_path.split('/')[-1][:-2] # Cut .c
-    params = maze_gen.get_params_from_maze(maze,'_') # Give a non-empty SMT path to trigger smt conversion  
+    params = maze_gen.get_params_from_maze(maze,smt_path=smt_path)  
+    return params
+
+def set_fake_params(params):
     params['t'] = params['t'].replace('storm', '')
     params['t'] = params['t'].replace('__', '_')
     if params['t'].endswith('_'):
@@ -87,20 +96,18 @@ def get_params(maze_path):
         params['m'] = 0
     return params
 
+
 def load(argv):
     maze = argv[0]
     tool = argv[1]
     if len(argv) == 6:
         variant = argv[2]
-        mutant = argv[3]
+        seeddir = argv[3]
         outdir = argv[4]
         timeout = argv[5]
-        #gen = argv[6]
     else:
         variant = ''
-        mutant = argv[2]
+        seeddir = argv[2]
         outdir = argv[3]
         timeout = argv[4]
-        #gen = argv[5]
-    memory = 4
-    main(maze, tool, variant, memory, mutant,outdir,timeout)
+    main(maze, tool, variant,seeddir,outdir,timeout)
